Minor changes to scrolly.
Trying to get visualizations to show up.  There are some problems
because people created them in divs that already were sized.  I need
to work out this problems so that the visualizations all create
themselves at the same size.  There seems to be a problem if the width
changes for some transitions.  It will go into an infinite loop of
back and forth under the right conditions.