Fixed problems and extend RidersPerMinute.
Fixed a few problems with RidersPerMinute.  Added RidersPerMinuteDay
to be able to keep track of the days for riders that go over.  Also, I
created a Spark query to get all rides for a date range and then
partition it into cities.  That way, I can use the new
RidersPerMinuteDay code to generate separate CSVs for each day.