Fix layout issues for graphs page.
Fix layout issues for graphs second page.  Shrunk top graph to half of
the figure.  Added a title to the stacked bar graph and made the text
for the cities match that in the bar graph.  There are still problems
with the tooltop being behind the graph.  I'm working on this.