Rpm style nits (#54)
* Adjusted some styles; changed date formats.

Tinkered with padding so spacing is more consistent.  Changed date
formats to be more constent.

* Change tooltip to not waste space on right.  Adjust margins.

Fix the tooltip so that it switches when the user gets too far to the
right.  That way we can save space by having smaller more consistent
margins on the right side.

* Fix something requested in a review.

Remove a debug line requested in a reviewer comment.

* Update style.css

fix: removing merge fragment

Co-authored-by: Richard Morello <richard.morello@gmail.com>
Co-authored-by: hewness <gordonhew@gmail.com>